+++ src/decorators/index.ts
@@ -5,3 +5,4 @@ export * from './componentDisplayFlex/componentDisplayFlex.decorator';
 export * from './componentHostStyle/componentHostStyle.decorator';
 export * from './scrollableContent/scrollableContent.decorator';
 export * from './typeProvider/typeProvider.decorator';
+export * from './withScrollableCssClass/withScrollableCssClass.decoator';