+++ src/types/host/index.ts
@@ -2,3 +2,8 @@
  * Enables animation of children on components
  */
 export const AnimateChildrenHostFeature = {'[@animateChildren]': 'true'};
+
+/**
+ * Adds css class 'scrollable' to host element
+ */
+export const ScrollableContentHostFeature = {'[class.scrollable]': 'true'};